chore: Refactor dependency Map code (#14944)
