+++ app/client/src/workers/DataTreeEvaluator/test.ts
@@ -5,6 +5,7 @@ import { DataTreeDiff } from "workers/evaluationUtils";
 import { ALL_WIDGETS_AND_CONFIG } from "utils/WidgetRegistry";
 import { arrayAccessorCyclicDependency } from "./mockData/ArrayAccessorTree";
 import { nestedArrayAccessorCyclicDependency } from "./mockData/NestedArrayAccessorTree";
+import { updateDependencyMap } from "workers/DependencyMap";
 import { parseJSActions } from "workers/JSObject";
 
 const widgetConfigMap = {};
@@ -155,10 +156,11 @@ describe("DataTreeEvaluator", () => {
           event: "EDIT",
         },
       ];
-      dataTreeEvaluator.updateDependencyMap(
-        translatedDiffs as Array<DataTreeDiff>,
-        dataTreeEvaluator.oldUnEvalTree,
-      );
+      updateDependencyMap({
+        dataTreeEvalRef: dataTreeEvaluator,
+        translatedDiffs: translatedDiffs as Array<DataTreeDiff>,
+        unEvalDataTree: dataTreeEvaluator.oldUnEvalTree,
+      });
 
       expect(dataTreeEvaluator.dependencyMap).toStrictEqual({
         "Button2.text": [],
@@ -177,10 +179,11 @@ describe("DataTreeEvaluator", () => {
           event: "EDIT",
         },
       ];
-      dataTreeEvaluator.updateDependencyMap(
-        translatedDiffs as Array<DataTreeDiff>,
-        dataTreeEvaluator.oldUnEvalTree,
-      );
+      updateDependencyMap({
+        dataTreeEvalRef: dataTreeEvaluator,
+        translatedDiffs: translatedDiffs as Array<DataTreeDiff>,
+        unEvalDataTree: dataTreeEvaluator.oldUnEvalTree,
+      });
 
       expect(dataTreeEvaluator.dependencyMap).toStrictEqual({
         Button2: ["Button2.text"],
